Fixed critical vulnerability
I have suspected for some time that it is possible to hack my site, now I tried to do it and I succeeded. The trick is that in the input field you give, for example,  (two dots), which means the parent directory for Linux systems. And that's how you can escape the  directory and replace **any** file on the server.